Sync up common-modules/virtual-device with android-4.19-stable
To these commits in the android-4.19-stable branch:

cuttlefish-modules: ad139b498ba98fcd1f66953ea87d0fe960bee96b

There is no android-4.19-stable branch for goldfish, so it retains
its android12-5.4 code.

Bug: 166154615

+++ virtio_gpu/virtgpu_fence.c
@@ -23,8 +23,8 @@
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
+#include <drm/drmP.h>
 #include <trace/events/dma_fence.h>
-
 #include "virtgpu_drv.h"
 
 static const char *virtio_get_driver_name(struct dma_fence *f)
@@ -48,7 +48,7 @@ bool virtio_fence_signaled(struct dma_fence *f)
 
 static void virtio_fence_value_str(struct dma_fence *f, char *str, int size)
 {
-	snprintf(str, size, "%llu", f->seqno);
+	snprintf(str, size, "%llu", (long long unsigned int) f->seqno);
 }
 
 static void virtio_timeline_value_str(struct dma_fence *f, char *str, int size)
